requirements.txt
scapy>=2.5.0
watchdog>=3.0.0
slack-sdk>=3.21.0
python-dotenv>=1.0.0
requests>=2.31.0